Almost implemented assembly unpack for bit sizes 8-16 as well.
The decoding procedure (unpack_internal_bmi2_9_16) is slightly deficient
in its reading pattern (straightforward port from the 0_8 version would
have required reading with 0.5 byte offset for odd bitsizes. This can be
fixed up and is marked for this but I don't have time to do it now).

Anyhow, the new version is commented out right now, but should be also
some faster (385 Mints/s for the version with just 0_8 vs
400+ Mints/s for 0_8 + 9_16)

Some other changes:
- changed to unaligned writes, so no output buffer alignment guarantee
  is needed. No measurable performance change observed.
- remove the valloc() for that reason. Tested with addresses that are
  definitely unaligned and the performance is the same.